Fixing SBSA MP usage (#4)
The existing TFA will fault the secondary cores when placing the TFA image in secure flash. This was because when the stmm image is trying to communicate with the secure flash for UEFI variable support, the flash will change mode to write. In the meantime, the secondary cores looping at `wfe` will be woken up which at this point the flash cannot be read.

This change will bring the secondary cores to loop at BL31 instead of BL1 which will remove the execute on flash dependency to avoid such issue.

Known drawback: The context losing sleep will be broken because the core cannot fall back to the context losing state once this is brought up BL31 context.